+++ spec/chris_lib/chris_math_spec.rb
@@ -1,6 +1,54 @@
 # encoding: utf-8
 require 'spec_helper'
 require 'pp'
+include Math
+require 'quaternion'
+
+describe 'Quaternion extensions' do
+	describe 'quaternion.round(n)' do
+	  let!(:q) { Quaternion.new(1.003, -0.8345, 0.3456, 0.1345) }
+	  it { expect(q.round(2)).to eq Quaternion.new(1.00, -0.83, 0.35, 0.13) }
+	end
+
+	describe 'Quaternion.identity' do
+	  it { expect(Quaternion.identity).to eq Quaternion.new(1.0, 0.0, 0.0, 0.0) }
+	end
+
+	describe 'Quaternion.zero' do
+	  it { expect(Quaternion.zero).to eq Quaternion.new(0.0, 0.0, 0.0, 0.0) }
+	end
+
+	describe 'Quaternion initialization with brackets' do
+	  it { expect(Quaternion[1.0, 2.0, 3.0, 4.0]).to eq Quaternion.new(1.0, 2.0, 3.0, 4.0) }
+	end
+end
+
+describe 'Float degree-radian conversions' do
+  let!(:rad0) { PI / 6 }
+  let!(:rad1) { 58.316 / 180.0 * PI }
+  let!(:rad2) { 45 / 180.0 * PI }
+  describe 'Float.to_deg(n)' do
+    it { expect(rad0.to_deg).to be_within(1e-9).of  30.0 }
+    it { expect(rad1.to_deg(2)).to eq 58.32 }
+  end
+  describe 'Float.to_rad(n)' do
+    it { expect(30.0.to_rad).to be_within(1e-9).of rad0 }
+    it { expect(30.0.to_rad(3)).to eq 0.524 }
+  end
+
+  describe 'Array degree-radian conversions' do
+    describe 'Array.to_rad(n)' do
+      let!(:ary) { [[rad0, rad1], [rad0, []]]  }
+      it {expect(ary.to_deg.round(9)).to eq  [[30.0, 58.316], [30.0, []]] }
+      it {expect(ary.to_deg(2)).to eq  [[30.0, 58.32], [30.0, []]] }
+    end
+    describe 'Array.to_rad(n)' do
+      let!(:ary) { [[30.0, 45.0], [30.0, []]] }
+      it { expect(ary.to_rad.round(9)).to eq [[rad0.round(9), rad2.round(9)], [rad0.round(9), []]] }
+      it { expect(ary.to_rad.round(3)).to eq [[0.524, 0.785], [0.524, []]] }
+    end
+  end
+end
 
 describe 'Array.round(n)' do
   let!(:a) { [1.003, -0.8345, 0.3456, 0.1345] }